+++ src/main/java/com/example/be_hotel/service/UserService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface UserService {
     void addUser(User user);
     List<User> findAllUser();
-    User getUser(Integer id);
-    void updateUser(Integer id, User user);
-    void deleteUser(Integer id);
+    User getUser(Long id);
+    void updateUser(Long id, User user);
+    void deleteUser(Long id);
 }
